refactor: reorganize project files
