--- vl-main/src/desktop/mod.rs
@@ -1,2 +1,0 @@
-#[cfg(target_os = "linux")]
-pub mod linux;
