+++ vl-main/src/lib.rs
@@ -3,8 +3,7 @@
 #![feature(core_intrinsics)]
 
 use bevy::prelude::*;
-pub mod base_modules;
-pub mod desktop;
+pub mod modules;
 pub mod ui;
 
 #[bevy_main]
