+++ vl-main/src/main.rs
@@ -5,8 +5,7 @@
 #![feature(core_intrinsics)]
 
 use bevy::prelude::bevy_main;
-pub mod base_modules;
-pub mod desktop;
+pub mod modules;
 pub mod ui;
 
 #[bevy_main]
